+++ game-client/gameplay/scenes/testing.ts
@@ -1,5 +1,6 @@
 import { Scene } from '../util/phaserExtensions'
 import { Math, Tilemaps } from 'phaser'
+import { Player } from '../objects/player'
 
 const testMap = [
   [1, 0, 0, 1, 0, 3, 3, 0, 0, 2, 2, 1, 0, 5, 5, 1, 3, 1, 1, 1],
@@ -61,17 +62,11 @@ const scene = new Scene(sceneConfig)
 
 let tileMap: Tilemaps.Tilemap
 
-let player: Phaser.Types.Physics.Arcade.SpriteWithDynamicBody
+let player: Player
 const playerMovement = { x: 0, y: 0, direction: 'down' }
 let cursors: Phaser.Types.Input.Keyboard.CursorKeys
 
-const Directions = ['up', 'right', 'down', 'left']
-const WALK_LOOKUP: number[][] = [
-  [0, 1, 2, 1], // 0: up
-  [3, 4, 5, 4], // 1: right
-  [6, 7, 8, 7], // 2: down
-  [9, 10, 11, 10], // 3: left
-]
+
 
 function setPlayerMovement() {
   const up = cursors.up.isDown ? -1 : 0
@@ -125,33 +120,10 @@ scene.create = () => {
   const tileMapLayer = tileMap.createLayer(0, tiles, 0, 0)
   tileMapLayer.setCollision(0)
   tileMapLayer.setCollision(2)
-  // player
-  player = scene.physics.add.sprite(32 * 6, 32, 'TEST1.healer_f', 4)
-  for (const num in WALK_LOOKUP) {
-    const key: string = Directions[num]
-    player.anims.create({
-      key: `walk_${key}`,
-      frames: player.anims.generateFrameNumbers('TEST1.healer_f', {
-        frames: WALK_LOOKUP[num],
-      }),
-      frameRate: 3,
-      repeat: -1,
-    })
-    player.anims.create({
-      key: `idle_${key}`,
-      frames: player.anims.generateFrameNumbers('TEST1.healer_f', {
-        frames: [WALK_LOOKUP[num][1]],
-      }),
-      frameRate: 3,
-      repeat: -1,
-    })
-  }
+  player = new Player(32 * 6, 32, 0, scene, 'TEST1.healer_f')
   // collisions
   tileMap.setCollisionByProperty({collision: true}, true, true, tileMapLayer)
-  scene.physics.add.collider(player, tileMapLayer)
-  player.setCollideWorldBounds(true)
-  // init player anim
-  player.anims.play('idle_down', true)
+  scene.physics.add.collider(player.sprite, tileMapLayer)
 }
 
 scene.update = () => {
@@ -159,12 +131,12 @@ scene.update = () => {
   const animType =
     playerMovement.x !== 0 || playerMovement.y !== 0 ? 'walk' : 'idle'
   const animation = `${animType}_${playerMovement.direction}`
-  if (animation !== player.anims.currentAnim.key) {
-    player.anims.play(animation, true)
+  if (animation !== player.sprite.anims.currentAnim.key) {
+    player.sprite.anims.play(animation, true)
   }
 
   const inputVector: Math.Vector2 = new Math.Vector2(playerMovement).normalize()
-  player.setVelocity(inputVector.x * 100, inputVector.y * 100)
+  player.sprite.setVelocity(inputVector.x * 100, inputVector.y * 100)
 }
 
 export default scene
